+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Linter's execution action can be set by the `phpCodeSniffer.lintAction` option.
 - Command `phpCodeSniffer.cancelProcessing` to cancel all active processing.
 
 ## [1.3.0] - 2021-05-24
@@ -24,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.0] - 2021-03-04
 ### Added
-- Ignore diagnostics for files using `phpCodeSniffer.ignorePatterns` option.
+- Ignore diagnostics for files using the `phpCodeSniffer.ignorePatterns` option.
 
 ### Fixed
 - "Ignore * for this line" action should be present for all diagnostics.